script.py
import calendar
import dateutil.parser
import timeout_decorator
import bz2
import beanstalkc
import collections
import functools
import time
from urllib2 import HTTPError
from heapq import heappush, heappop
from itertools import count
import math
from operator import mul
import os
from graph_tool.all import *
from ripe.atlas.sagan import Result
import radix
from ripe.atlas.cousteau import Probe, Measurement
import pdb
import urllib2
import urllib
from ripe.atlas.cousteau import (
    Traceroute,
    AtlasSource,
    AtlasCreateRequest,
    ProbeRequest,
    AtlasResultsRequest
)
import random
import json
import csv
import glob

import mkit.ripeatlas.probes as prb
import mkit.inference.ixp as ixp
import mkit.inference.ip_to_asn as ip2asn

# local consts and settings (consts.py and settings.py)
from consts import *
from settings import *

g = Graph()
print "test"
